feat: move encryption and utils out of client
