+++ autonomi/src/client/high_level/files/fs_public.rs
@@ -68,7 +68,8 @@ impl Client {
         info!("Uploading directory: {dir_path:?}");
 
         // encrypt
-        let encryption_results = self.encrypt_directory_files(dir_path, true).await?;
+        let encryption_results =
+            crate::self_encryption::encrypt_directory_files(dir_path, true).await?;
         let mut chunk_iterators = vec![];
         for encryption_result in encryption_results {
             match encryption_result {
