+++ autonomi/src/self_encryption/mod.rs
@@ -1,0 +1,14 @@
+// Copyright 2025 MaidSafe.net limited.
+//
+// This SAFE Network Software is licensed to you under The General Public License (GPL), version 3.
+// Unless required by applicable law or agreed to in writing, the SAFE Network Software distributed
+// under the GPL Licence is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied. Please review the Licences for the specific language governing
+// permissions and limitations relating to use of the SAFE Network Software.
+
+mod memory_encryption;
+mod stream_encryption;
+
+pub use memory_encryption::{DataMapLevel, Error, encrypt};
+pub use self_encryption::MAX_CHUNK_SIZE;
+pub(crate) use stream_encryption::{EncryptionStream, encrypt_directory_files};
